Added upload progress screen
