+++ desktop/src/renderer/src/assets/base.css
@@ -30,6 +30,8 @@
 
 	--modal-border: #d8d8d8;
 
+	--upload-progress-bar: #5595d9;
+
 	--progress-complete: #42814a;
 	--progress-incomplete: #f8bb47;
 	--progress-cancelled: #fe3f35;
