+++ desktop/src/renderer/src/components/transfer/lan-views/index.ts
@@ -2,4 +2,4 @@ export * from "./LanUploadFileView";
 export * from "./LanUploadTransferFormView";
 export * from "./LanSubmissionAgreementView";
 export * from "./LanUploadView";
-export * from "./LanCompletionView";
+export * from "./LanFinishView";
